Release v1.0.6: fix CI flaky test, add CLI commands, Python binding
Bug fixes:
- Encoder: deep-copy string/blob value data to prevent use-after-return
  when JSON encoder passes stack-allocated buffers (root cause of CI
  flaky test on ubuntu-latest/clang)
- JSON decoder: fix heap-buffer-overflow in segment dedup
- Bitstream: fix UB in zigzag encode (left-shift of negative int64_t)

New features:
- CLI tool: trp encode/decode/validate commands with stdin and -o support
- Native Python binding with 70 tests
- Complex JSON example (json_complex.c)
- 3 new C test files, expanded edge-case coverage (~400 total tests)
- README: CI/coverage/license badges, roadmap, updated bindings table

+++ src/bitstream/bitstream_varint.c
@@ -80,7 +80,7 @@ tp_result tp_bs_write_varint_s(tp_bitstream_writer *w, int64_t value)
     if (!w)
         return TP_ERR_INVALID_PARAM;
 
-    /* Zigzag encode */
-    uint64_t raw = (uint64_t)((value << 1) ^ (value >> 63));
+    /* Zigzag encode: cast to unsigned before left shift to avoid UB */
+    uint64_t raw = ((uint64_t)value << 1) ^ (uint64_t)(value >> 63);
     return tp_bs_write_varint_u(w, raw);
 }
